Dodan mobile layout i za my profile section
