+++ tv_shows_app/src/components/features/auth/LoginForm/layouts/LoginForm.desktop.tsx
@@ -110,6 +110,7 @@ export const LoginFormDesktop = () => {
             loadingText="Logging in"
             type="submit"
             variant="default"
+            textStyle="button"
           >
             LOG IN
           </Button>
